remove unneccessary empty argument list

+++ src/main/scala/edu/ie3/powerFactory2psdm/converter/Transformer2WConverter.scala
@@ -43,7 +43,7 @@ object Transformer2WConverter extends LazyLogging {
       nodeB,
       1,
       trafoType,
-      input.tapPos.toInt,
+      input.tapPos,
       autotap
     )
   }
